Chore #213 : relocate component
