+++ feature/my/src/main/java/com/hankki/feature/my/myjogbodetail/component/JogboShareButton.kt
@@ -1,4 +1,4 @@
-package com.hankki.feature.my.component
+package com.hankki.feature.my.myjogbodetail.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
