+++ feature/my/src/main/java/com/hankki/feature/my/mypage/component/ButtonWithImageAndBorder.kt
@@ -1,4 +1,4 @@
-package com.hankki.feature.my.component
+package com.hankki.feature.my.mypage.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -15,11 +15,11 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.ColorFilter
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.hankki.core.designsystem.R
 import com.hankki.core.designsystem.theme.Gray200
 import com.hankki.core.designsystem.theme.Gray400
 import com.hankki.core.designsystem.theme.Gray50
@@ -69,7 +69,7 @@ fun ButtonWithImageAndBorder(
 fun ButtonPrev() {
     HankkijogboTheme {
         ButtonWithImageAndBorder(
-            image = com.hankki.core.designsystem.R.drawable.ic_good,
+            image = R.drawable.ic_good,
             description = "test",
             modifier = Modifier
         )
